refactor: moving getReleaseDefaults to utils
