+++ packages/sanity/src/core/releases/tool/overview/ReleasesOverview.tsx
@@ -33,8 +33,8 @@ import {useArchivedReleases} from '../../store/useArchivedReleases'
 import {useReleaseOperations} from '../../store/useReleaseOperations'
 import {useReleasePermissions} from '../../store/useReleasePermissions'
 import {type ReleasesMetadata, useReleasesMetadata} from '../../store/useReleasesMetadata'
-import {getReleaseDefaults} from '../../util/const'
 import {getReleaseTone} from '../../util/getReleaseTone'
+import {getReleaseDefaults} from '../../util/util'
 import {ReleaseMenuButton} from '../components/ReleaseMenuButton/ReleaseMenuButton'
 import {Table, type TableRowProps} from '../components/Table/Table'
 import {type TableSort} from '../components/Table/TableProvider'
