+++ packages/sanity/src/core/releases/util/const.ts
@@ -1,8 +1,3 @@
-/*  TEMPORARY  DUMMY DATA */
-
-import {type EditableReleaseDocument} from '../store/types'
-import {createReleaseId} from './createReleaseId'
-
 /**
  * @internal
  */
@@ -15,12 +10,3 @@ export const PUBLISHED = 'published' as const
 export const DEFAULT_RELEASE_TYPE = 'asap'
 
 export const ARCHIVED_RELEASE_STATES = ['archived', 'published']
-
-export const getReleaseDefaults: () => EditableReleaseDocument = () => ({
-  _id: createReleaseId(),
-  metadata: {
-    title: '',
-    description: '',
-    releaseType: DEFAULT_RELEASE_TYPE,
-  },
-})
